cast tp precision

+++ unifold/train/train_config.py
@@ -51,7 +51,7 @@ train_config = ConfigDict({
         'load_dir': '../out/unifold',
         # Training precision, generally in ['fp32', 'bf16'].
         # Set for mixed precision training.
-        'precision': 'fp16',
+        'precision': 'bf16',
         # Max queue size. Specifies the queue size of the pre-processed
         # batches. Generally has little impact on code efficiency.
         'max_queue_size': 128,
